+++ pkg/mads/types.go
@@ -1,0 +1,5 @@
+package mads
+
+const (
+	MonitoringAssignmentType = "type.googleapis.com/kuma.observability.v1alpha1.MonitoringAssignment"
+)